moved state from App to ThreadItemContainer along with functions
