+++ client/src/components/ThreadItemContainer/ThreadItemContainer.js
@@ -1,24 +1,85 @@
 
 import React, { Component } from 'react';
 import ThreadItem from '../ThreadItem/ThreadItem';
+import axios from 'axios';
+import update from 'immutability-helper';
 
-function ThreadItemContainer(props){
-  return (
-    <div className="col m8">
-      <div className="card-panel white z-depth-4"> 
-        <div className="card-action">
-          <h3>Threads</h3> 
+class ThreadItemContainer extends Component {
+  // has to be switched back to a class Component
+  // this container will contain the two methods
+
+  constructor(props){
+    super(props);
+
+    this.state = {
+      threads: []
+    };
+
+    // without these, instances don't have these methods
+    this.upvoteCount = this.upvoteCount.bind(this);
+    this.downvoteCount = this.downvoteCount.bind(this);
+  }
+
+  // should have two methods here where I can update the counts
+  upvoteCount(idx) {
+    // update the thread's count
+    console.log('updating the count na mean');
+    const updatedCount = update(this.state.threads[idx], {votes: {$apply: (x) => x+1}}); 
+    let copy = this.state.threads.slice();
+    copy[idx] = updatedCount;
+
+    this.setState({threads: copy});
+  }
+
+  downvoteCount(idx) {
+    // update the thread's count
+    const updatedCount = update(this.state.threads[idx], {votes: {$apply: (x) => x-1}}); 
+    let copy = this.state.threads.slice();
+    copy[idx] = updatedCount;
+
+    this.setState({threads: copy});
+  }
+
+  componentDidMount() {
+    // const url = 'http://rest.learncode.academy/api/am/friends'; 
+
+    const url = '/posts/'; 
+    axios.get(url)
+      .then(res => {
+        console.log(res.data);
+        console.log(res.data[0]._id);
+        this.setState({threads: res.data})
+      });
+
+    // mongo jobquest --eval "db.dropDatabase()" => to reset DB
+
+    // const dummy = {title: 'new stuff brahh', thread: 'Nonsense bahh'};
+    // axios.post('/posts/', dummy)
+    // .then((res) => {console.log('the res is, ', res)})
+    // .catch(err => {
+    //   console.log(err);
+    // });
+  }
+
+  render(){
+    return (
+      <div className="col m8">
+        <div className="card-panel white z-depth-4"> 
+          <div className="card-action">
+            <h3>Threads</h3> 
+          </div>
+
+          {this.state.threads ? 
+            this.state.threads.map((thread, idx) => {
+              return <ThreadItem key={idx} data={thread} upvote={() => this.upvoteCount(idx)}
+                                                         downvote={() => this.downvoteCount(idx)}/>
+            })
+            : null}
         </div>
-
-        {props.items ? 
-          props.items.map((thread, idx) => {
-            return <ThreadItem key={idx} data={thread} upvote={() => props.upvote(idx)}
-                                                       downvote={() => props.downvote(idx)}/>
-          })
-          : null}
       </div>
-    </div>
-  );
+    );
+  }
+
 }
 
 ThreadItemContainer.PropTypes = {
